+++ src/build-no-otr.js
@@ -1,0 +1,31 @@
+({
+    baseUrl: "../",
+    name: "components/almond/almond.js",
+    out: "../builds/converse-no-otr.min.js",
+    include: ['main'],
+    paths: {
+        "jquery": "components/jquery/jquery",
+        "jed": "components/jed/jed",
+        "locales": "locale/locales",
+        "af": "locale/af/LC_MESSAGES/af",
+        "de": "locale/de/LC_MESSAGES/de",
+        "en": "locale/en/LC_MESSAGES/en",
+        "es": "locale/es/LC_MESSAGES/es",
+        "fr": "locale/fr/LC_MESSAGES/fr",
+        "hu": "locale/hu/LC_MESSAGES/hu",
+        "it": "locale/it/LC_MESSAGES/it",
+        "nl": "locale/nl/LC_MESSAGES/nl",
+        "pt_BR": "locale/pt_BR/LC_MESSAGES/pt_BR", 
+        "ru": "locale/ru/LC_MESSAGES/ru",
+        "jquery.tinysort": "components/tinysort/src/jquery.tinysort",
+        "underscore": "components/underscore/underscore",
+        "backbone": "components/backbone/backbone",
+        "backbone.localStorage": "components/backbone.localStorage/backbone.localStorage",
+        "strophe": "components/strophe/strophe",
+        "strophe.muc": "components/strophe.muc/index",
+        "strophe.roster": "components/strophe.roster/index",
+        "strophe.vcard": "components/strophe.vcard/index",
+        "strophe.disco": "components/strophe.disco/index",
+        "converse-dependencies": "src/deps-no-otr"
+    }
+})