+++ src/deps-full.js
@@ -1,0 +1,13 @@
+define("converse-dependencies", [
+    "otr",
+    "locales",
+    "backbone.localStorage",
+    "jquery.tinysort",
+    "strophe",
+    "strophe.muc",
+    "strophe.roster",
+    "strophe.vcard",
+    "strophe.disco"
+], function(otr) {
+    return otr;
+});